JS: refactor
