+++ js/js.go
@@ -2,8 +2,7 @@
 package js
 
 // TODO: remove dead code (if(false) or after return/throw/break/continue), difficulty with var/func decls
-// TODO: move var declaration or expr statement into for loop init (var only if for has var decl)
-// TODO: don't minify variable names in with statement, what todo with eval? Don't minify any variable name?
+// TODO: don't minify variable names when code has with statement or eval call
 
 import (
 	"bytes"
@@ -367,237 +366,6 @@ func (m *jsMinifier) minifyStmt(i js.IStmt) {
 	}
 }
 
-func groupExpr(i js.IExpr, prec js.OpPrec) js.IExpr {
-	if exprPrec(i) < prec {
-		return &js.GroupExpr{i}
-	}
-	return i
-}
-
-func condExpr(x, y, z js.IExpr) js.IExpr {
-	return &js.CondExpr{groupExpr(x, js.OpCoalesce), groupExpr(y, js.OpAssign), groupExpr(z, js.OpAssign)}
-}
-
-func (m *jsMinifier) optimizeStmt(i js.IStmt) js.IStmt {
-	// convert if/else into expression statement, and optimize blocks
-	if ifStmt, ok := i.(*js.IfStmt); ok {
-		hasIf := !m.isEmptyStmt(ifStmt.Body)
-		hasElse := !m.isEmptyStmt(ifStmt.Else)
-		if unaryExpr, ok := ifStmt.Cond.(*js.UnaryExpr); ok && unaryExpr.Op == js.NotToken && hasElse {
-			ifStmt.Cond = unaryExpr.X
-			ifStmt.Body, ifStmt.Else = ifStmt.Else, ifStmt.Body
-			hasIf, hasElse = hasElse, hasIf
-		}
-		if !hasIf && !hasElse {
-			return &js.ExprStmt{ifStmt.Cond}
-		} else if hasIf && !hasElse {
-			ifStmt.Body = m.optimizeStmt(ifStmt.Body)
-			if X, isExprBody := ifStmt.Body.(*js.ExprStmt); isExprBody {
-				if unaryExpr, ok := ifStmt.Cond.(*js.UnaryExpr); ok && unaryExpr.Op == js.NotToken {
-					left := groupExpr(unaryExpr.X, binaryLeftPrecMap[js.OrToken])
-					right := groupExpr(X.Value, binaryRightPrecMap[js.OrToken])
-					return &js.ExprStmt{&js.BinaryExpr{js.OrToken, left, right}}
-				} else {
-					left := groupExpr(ifStmt.Cond, binaryLeftPrecMap[js.AndToken])
-					right := groupExpr(X.Value, binaryRightPrecMap[js.AndToken])
-					return &js.ExprStmt{&js.BinaryExpr{js.AndToken, left, right}}
-				}
-			}
-		} else if !hasIf && hasElse {
-			ifStmt.Else = m.optimizeStmt(ifStmt.Else)
-			if X, isExprElse := ifStmt.Else.(*js.ExprStmt); isExprElse {
-				left := groupExpr(ifStmt.Cond, binaryLeftPrecMap[js.OrToken])
-				right := groupExpr(X.Value, binaryRightPrecMap[js.OrToken])
-				return &js.ExprStmt{&js.BinaryExpr{js.OrToken, left, right}}
-			}
-		} else if hasIf && hasElse {
-			ifStmt.Body = m.optimizeStmt(ifStmt.Body)
-			ifStmt.Else = m.optimizeStmt(ifStmt.Else)
-			XExpr, isExprBody := ifStmt.Body.(*js.ExprStmt)
-			YExpr, isExprElse := ifStmt.Else.(*js.ExprStmt)
-			if isExprBody && isExprElse {
-				return &js.ExprStmt{condExpr(ifStmt.Cond, XExpr.Value, YExpr.Value)}
-			}
-			XReturn, isReturnBody := ifStmt.Body.(*js.ReturnStmt)
-			YReturn, isReturnElse := ifStmt.Else.(*js.ReturnStmt)
-			if isReturnBody && isReturnElse {
-				if XReturn.Value == nil && YReturn.Value == nil {
-					return &js.ReturnStmt{&js.BinaryExpr{js.CommaToken, ifStmt.Cond, &js.UnaryExpr{js.VoidToken, &js.LiteralExpr{js.NumericToken, zeroBytes}}}}
-				} else if XReturn.Value != nil && YReturn.Value != nil {
-					return &js.ReturnStmt{condExpr(ifStmt.Cond, XReturn.Value, YReturn.Value)}
-				}
-				return ifStmt
-			}
-			XThrow, isThrowBody := ifStmt.Body.(*js.ThrowStmt)
-			YThrow, isThrowElse := ifStmt.Else.(*js.ThrowStmt)
-			if isThrowBody && isThrowElse {
-				return &js.ThrowStmt{condExpr(ifStmt.Cond, XThrow.Value, YThrow.Value)}
-			}
-		}
-	} else if decl, ok := i.(*js.VarDecl); ok && m.varsHoisted != nil && decl != m.varsHoisted {
-		for _, item := range decl.List {
-			if item.Default != nil {
-				return &js.ExprStmt{decl}
-			}
-		}
-		return &js.EmptyStmt{}
-	} else if blockStmt, ok := i.(*js.BlockStmt); ok {
-		// merge body and remove braces if it is not a lexical declaration
-		blockStmt.List = m.optimizeStmtList(blockStmt.List, defaultBlock)
-		if len(blockStmt.List) == 1 {
-			varDecl, isVarDecl := blockStmt.List[0].(*js.VarDecl)
-			_, isClassDecl := blockStmt.List[0].(*js.ClassDecl)
-			if !isClassDecl && (!isVarDecl || varDecl.TokenType == js.VarToken) {
-				return m.optimizeStmt(blockStmt.List[0])
-			}
-		} else if len(blockStmt.List) == 0 {
-			return &js.EmptyStmt{}
-		}
-		return blockStmt
-	}
-	return i
-}
-
-func (m *jsMinifier) flattenStmt(list []js.IStmt, i int) []js.IStmt {
-	if ifStmt, ok := list[i].(*js.IfStmt); ok && !m.isEmptyStmt(ifStmt.Else) && isFlowStmt(lastStmt(ifStmt.Body)) {
-		// if body ends in flow statement (return, throw, break, continue), so we can remove the else statement and put its body in the current scope
-		if blockStmt, ok := ifStmt.Else.(*js.BlockStmt); ok {
-			list = append(append(list[:i+1], blockStmt.List...), list[i+1:]...)
-		} else {
-			list = append(append(list[:i+1], ifStmt.Else), list[i+1:]...)
-		}
-		ifStmt.Else = nil
-	}
-	return list
-}
-
-func (m *jsMinifier) optimizeStmtList(list []js.IStmt, blockType blockType) []js.IStmt {
-	// merge expression statements as well as if/else statements followed by flow control statements
-	if len(list) == 0 {
-		return list
-	}
-	j := 0
-	list[0] = m.optimizeStmt(list[0])
-	if _, ok := list[0].(*js.EmptyStmt); ok {
-		j--
-	} else {
-		list = m.flattenStmt(list, 0)
-	}
-	for i, _ := range list[:len(list)-1] {
-		// probe at every i allowing one lookahead to i+1, write to position j <= i
-		j++
-		list[i+1] = m.optimizeStmt(list[i+1])
-		if _, ok := list[i+1].(*js.EmptyStmt); ok {
-			j--
-			continue
-		} else {
-			list = m.flattenStmt(list, i+1)
-		}
-
-		// merge expression statements with expression, return, and throw statements
-		if left, ok := list[i].(*js.ExprStmt); ok {
-			if right, ok := list[i+1].(*js.ExprStmt); ok {
-				right.Value = &js.BinaryExpr{js.CommaToken, left.Value, right.Value}
-				j--
-			} else if returnStmt, ok := list[i+1].(*js.ReturnStmt); ok && returnStmt.Value != nil {
-				returnStmt.Value = &js.BinaryExpr{js.CommaToken, left.Value, returnStmt.Value}
-				j--
-			} else if throwStmt, ok := list[i+1].(*js.ThrowStmt); ok {
-				throwStmt.Value = &js.BinaryExpr{js.CommaToken, left.Value, throwStmt.Value}
-				j--
-			} else if forStmt, ok := list[i+1].(*js.ForStmt); ok {
-				if forStmt.Init == nil {
-					forStmt.Init = left.Value
-					j--
-				} else if _, ok := forStmt.Init.(*js.VarDecl); !ok {
-					forStmt.Init = &js.BinaryExpr{js.CommaToken, left.Value, forStmt.Init}
-					j--
-				}
-			} else if whileStmt, ok := list[i+1].(*js.WhileStmt); ok {
-				var body js.BlockStmt
-				if blockStmt, ok := whileStmt.Body.(*js.BlockStmt); ok {
-					body = *blockStmt
-				} else {
-					body.List = []js.IStmt{whileStmt.Body}
-				}
-				list[i+1] = &js.ForStmt{left.Value, whileStmt.Cond, nil, body}
-				j--
-			} else if switchStmt, ok := list[i+1].(*js.SwitchStmt); ok {
-				switchStmt.Init = &js.BinaryExpr{js.CommaToken, left.Value, switchStmt.Init}
-				j--
-			} else if withStmt, ok := list[i+1].(*js.WithStmt); ok {
-				withStmt.Cond = &js.BinaryExpr{js.CommaToken, left.Value, withStmt.Cond}
-				j--
-			} else if ifStmt, ok := list[i+1].(*js.IfStmt); ok {
-				ifStmt.Cond = &js.BinaryExpr{js.CommaToken, left.Value, ifStmt.Cond}
-				j--
-			}
-		} else if left, ok := list[i].(*js.VarDecl); ok && left.TokenType != js.VarToken {
-			// merge const, let declarations
-			if right, ok := list[i+1].(*js.VarDecl); ok && left.TokenType == right.TokenType {
-				right.List = append(left.List, right.List...)
-				j--
-			}
-		}
-		list[j] = list[i+1]
-
-		// merge if/else with return/throw when followed by return/throw
-		if 0 < j {
-			if ifStmt, ok := list[j-1].(*js.IfStmt); ok && m.isEmptyStmt(ifStmt.Body) != m.isEmptyStmt(ifStmt.Else) {
-				// either the if body is empty or the else body is empty. In case where both bodies have return/throw, we already rewrote that if statement to an return/throw statement
-				if returnStmt, ok := list[j].(*js.ReturnStmt); ok {
-					if returnStmt.Value == nil {
-						if left, ok := ifStmt.Body.(*js.ReturnStmt); ok && left.Value == nil {
-							list[j-1] = &js.ExprStmt{ifStmt.Cond}
-						} else if left, ok := ifStmt.Else.(*js.ReturnStmt); ok && left.Value == nil {
-							list[j-1] = &js.ExprStmt{ifStmt.Cond}
-						}
-					} else {
-						if left, ok := ifStmt.Body.(*js.ReturnStmt); ok && left.Value != nil {
-							returnStmt.Value = condExpr(ifStmt.Cond, left.Value, returnStmt.Value)
-							list[j-1] = returnStmt
-							j--
-						} else if left, ok := ifStmt.Else.(*js.ReturnStmt); ok && left.Value != nil {
-							returnStmt.Value = condExpr(ifStmt.Cond, returnStmt.Value, left.Value)
-							list[j-1] = returnStmt
-							j--
-						}
-					}
-				} else if throwStmt, ok := list[j].(*js.ThrowStmt); ok {
-					if left, ok := ifStmt.Body.(*js.ThrowStmt); ok {
-						throwStmt.Value = condExpr(ifStmt.Cond, left.Value, throwStmt.Value)
-						list[j-1] = throwStmt
-						j--
-					} else if left, ok := ifStmt.Else.(*js.ThrowStmt); ok {
-						throwStmt.Value = condExpr(ifStmt.Cond, throwStmt.Value, left.Value)
-						list[j-1] = throwStmt
-						j--
-					}
-				}
-			}
-		}
-	}
-
-	// remove superfluous return or continue
-	if 0 <= j {
-		if blockType == functionBlock {
-			if returnStmt, ok := list[j].(*js.ReturnStmt); ok {
-				if returnStmt.Value == nil || m.isUndefined(returnStmt.Value) {
-					j--
-				} else if binaryExpr, ok := returnStmt.Value.(*js.BinaryExpr); ok && binaryExpr.Op == js.CommaToken && m.isUndefined(binaryExpr.Y) {
-					// rewrite function f(){return a,void 0} => function f(){a}
-					list[j] = &js.ExprStmt{binaryExpr.X}
-				}
-			}
-		} else if blockType == iterationBlock {
-			if branchStmt, ok := list[j].(*js.BranchStmt); ok && branchStmt.Type == js.ContinueToken && branchStmt.Label == nil {
-				j--
-			}
-		}
-	}
-	return list[:j+1]
-}
-
 func (m *jsMinifier) minifyBlockStmt(stmt js.BlockStmt) {
 	m.write(openBraceBytes)
 	m.needsSemicolon = false
@@ -611,6 +379,8 @@ func (m *jsMinifier) minifyBlockStmt(stmt js.BlockStmt) {
 }
 
 func (m *jsMinifier) minifyBlockAsStmt(blockStmt *js.BlockStmt, blockType blockType) {
+	// minify block when statement is expected, i.e. semicolon if empty or remove braces for single statement
+	// assume we already renamed the scope
 	blockStmt.List = m.optimizeStmtList(blockStmt.List, blockType)
 	if 1 < len(blockStmt.List) {
 		m.minifyBlockStmt(*blockStmt)
@@ -623,6 +393,7 @@ func (m *jsMinifier) minifyBlockAsStmt(blockStmt *js.BlockStmt, blockType blockT
 }
 
 func (m *jsMinifier) minifyStmtOrBlock(i js.IStmt, blockType blockType) {
+	// minify stmt or a block
 	if blockStmt, ok := i.(*js.BlockStmt); ok {
 		m.renamer.renameScope(blockStmt.Scope)
 		m.minifyBlockAsStmt(blockStmt, blockType)
@@ -706,101 +477,6 @@ func (m *jsMinifier) minifyVarDecl(decl *js.VarDecl, onlyDefines bool) {
 			m.minifyBindingElement(item)
 		}
 	}
-}
-
-func (m *jsMinifier) hoistVars(body *js.BlockStmt) *js.VarDecl {
-	// Hoist all variable declarations in the current module/function scope to the top.
-	// If the first statement is a var declaration, expand it. Otherwise prepend a new var declaration.
-	// Except for the first var declaration, all others are converted to expressions. This is possible because an ArrayBindingPattern and ObjectBindingPattern can be converted to an ArrayLiteral or ObjectLiteral respectively, as they are supersets of the BindingPatterns.
-	parentVarsHoisted := m.varsHoisted
-	m.varsHoisted = nil
-	if 1 < body.Scope.NVarDecls {
-		iDefines := 0 // position past last variable definition in declaration
-		var decl *js.VarDecl
-		if varDecl, ok := body.List[0].(*js.VarDecl); ok && varDecl.TokenType == js.VarToken {
-			decl = varDecl
-		} else if forStmt, ok := body.List[0].(*js.ForStmt); ok && forStmt.Init != nil {
-			if varDecl, ok := forStmt.Init.(*js.VarDecl); ok && varDecl.TokenType == js.VarToken {
-				decl = varDecl
-			}
-		} else if whileStmt, ok := body.List[0].(*js.WhileStmt); ok {
-			decl = &js.VarDecl{js.VarToken, nil}
-			var forBody js.BlockStmt
-			if blockStmt, ok := whileStmt.Body.(*js.BlockStmt); ok {
-				forBody = *blockStmt
-			} else {
-				forBody.List = []js.IStmt{whileStmt.Body}
-			}
-			body.List[0] = &js.ForStmt{decl, whileStmt.Cond, nil, forBody}
-		}
-		if decl != nil {
-			// original declarations
-			vs := []*js.Var{}
-			for i, item := range decl.List {
-				if item.Default != nil {
-					iDefines = i + 1
-				}
-				vs = append(vs, bindingRefs(item.Binding)...)
-			}
-
-			// hoist other variable declarations in this function scope but don't initialize yet
-		DeclaredLoop:
-			for _, v := range body.Scope.Declared {
-				if v.Decl == js.VariableDecl {
-					for _, vdef := range vs {
-						if v == vdef {
-							continue DeclaredLoop
-						}
-					}
-					decl.List = append(decl.List, js.BindingElement{v, nil})
-				}
-			}
-		} else {
-			decl = &js.VarDecl{js.VarToken, nil}
-			for _, v := range body.Scope.Declared {
-				if v.Decl == js.VariableDecl {
-					decl.List = append(decl.List, js.BindingElement{v, nil})
-				}
-			}
-			body.List = append([]js.IStmt{decl}, body.List...)
-		}
-
-		// pull in assignments to variables into the declaration, e.g. var a;a=5  =>  var a=5
-		// sort in order of definitions
-		nMerged := 0
-	FindDefinitionsLoop:
-		for _, item := range body.List[1:] {
-			if exprStmt, ok := item.(*js.ExprStmt); ok {
-				if binaryExpr, ok := exprStmt.Value.(*js.BinaryExpr); ok && binaryExpr.Op == js.EqToken {
-					if v, ok := binaryExpr.X.(*js.Var); ok && v.Decl == js.VariableDecl {
-						if addDefinition(decl, iDefines, v, binaryExpr.Y) {
-							iDefines++
-							nMerged++
-							continue
-						}
-					}
-				}
-			} else if varDecl, ok := item.(*js.VarDecl); ok && varDecl.TokenType == js.VarToken {
-				for _, item := range varDecl.List {
-					if v, ok := item.Binding.(*js.Var); ok && item.Default != nil {
-						if addDefinition(decl, iDefines, v, item.Default) {
-							iDefines++
-							continue
-						}
-					}
-					break FindDefinitionsLoop // one of the declarations isn't a definition or can't be matched
-				}
-				nMerged++
-				continue // all variable declarations were matched, keep looking
-			}
-			break // not ExprStmt nor VarDecl
-		}
-		if 0 < nMerged {
-			body.List = append(body.List[:1], body.List[1+nMerged:]...)
-		}
-		m.varsHoisted = decl
-	}
-	return parentVarsHoisted
 }
 
 func (m *jsMinifier) minifyFuncDecl(decl js.FuncDecl, inExpr bool) {
@@ -1254,7 +930,7 @@ func (m *jsMinifier) minifyExpr(i js.IExpr, prec js.OpPrec) {
 			if expr.Generator {
 				m.write(starBytes)
 				m.minifyExpr(expr.X, js.OpAssign)
-			} else if v, ok := expr.X.(*js.Var); !ok || !bytes.Equal(v.Name(), undefinedBytes) { // TODO: only if not bound
+			} else if v, ok := expr.X.(*js.Var); !ok || !bytes.Equal(v.Name(), undefinedBytes) { // TODO: only if not defined
 				m.minifyExpr(expr.X, js.OpAssign)
 			}
 		}
@@ -1380,7 +1056,7 @@ func (m *jsMinifier) minifyExpr(i js.IExpr, prec js.OpPrec) {
 			m.minifyExpr(expr.Y, js.OpPrimary) // TemplateExpr or LiteralExpr
 		}
 	case *js.VarDecl:
-		m.minifyVarDecl(expr, true) // happens in when vars were hoisted
+		m.minifyVarDecl(expr, true) // happens in for statement or when vars were hoisted
 	case *js.FuncDecl:
 		if m.expectStmt {
 			m.write(notBytes)
